+++ graphbrain/parsers/parser.py
@@ -17,26 +17,6 @@ class Parser(object):
         # to be created by derived classes
         self.lang = None
 
-    def _post_process(self, edge):
-        raise NotImplementedError()
-
-    def _parse_token(self, token):
-        raise NotImplementedError()
-
-    def _before_parse_sentence(self):
-        raise NotImplementedError()
-
-    def _parse_sentence(self, sent):
-        raise NotImplementedError()
-
-    def _parse(self, text):
-        raise NotImplementedError()
-
-    def _resolve_corefs(self, parses):
-        # do nothing if not implemented in derived classes
-        for parse in parses:
-            parse['resolved_corefs'] = parse['main_edge']
-
     def parse(self, text):
         """Transforms the given text into hyperedges + aditional information.
         Returns a sequence of dictionaries, with one dictionary for each
@@ -56,3 +36,35 @@ class Parser(object):
         if self.resolve_corefs:
             self._resolve_corefs(parses)
         return parses
+
+    def atom_gender(self, atom):
+        raise NotImplementedError()
+
+    def atom_number(self, atom):
+        raise NotImplementedError()
+
+    def atom_person(self, atom):
+        raise NotImplementedError()
+
+    def atom_animacy(self, atom):
+        raise NotImplementedError()
+
+    def _post_process(self, edge):
+        raise NotImplementedError()
+
+    def _parse_token(self, token):
+        raise NotImplementedError()
+
+    def _before_parse_sentence(self):
+        raise NotImplementedError()
+
+    def _parse_sentence(self, sent):
+        raise NotImplementedError()
+
+    def _parse(self, text):
+        raise NotImplementedError()
+
+    def _resolve_corefs(self, parses):
+        # do nothing if not implemented in derived classes
+        for parse in parses:
+            parse['resolved_corefs'] = parse['main_edge']